WORK IN PROGRESS
Much of this is messy right now - just need to put this on github to
work on it with another pc with a gpu.

Attempting to use tanh for the image_classifier.py file and softmax
with two outputs for the image_classifier_softmax.py file. Softmax does
not yet work.